fix(desktop): use system apt and load the viewer after setup
