+++ ARCHITECTURE.md
@@ -803,7 +803,7 @@ graph LR
 
 A containerized assistant can serve an interactive desktop on demand. Desktop-only system packages and Google Chrome are installed only after the guardian clicks **Install desktop** in the modal. `GET /v1/desktop/setup` checks readiness without installing anything; `POST` starts one shared background installation. Both flat and assistant-scoped gateway paths require guardian authentication and proxy to a gateway-service-only runtime route. `assistant:self:desktop` sync invalidations refresh setup status as installation progresses, and a reopened modal or reconnected event stream refetches it. Older assistants returning 404 retain the direct streaming flow.
 
-`desktop-dependencies.ts` installs the X server, window manager, dock, compositor, clipboard bridge, terminal, fonts and Chrome libraries through apt without upgrading or removing installed packages. Google Chrome is an exact-version, SHA-256-verified download for Linux x64 or ARM64, extracted under the assistant's internal external-dependency directory. Extraction does not run Chrome package scripts, register its repository or change the system's default browser. Successful setup is recorded only after Chrome runs; readiness also checks the desktop binaries and X fonts so a recreated container offers setup again. Installation errors remain retryable. The base image carries no desktop-only packages or baked browser; existing browser-tool and PDF installations retain their own Playwright behavior.
+`desktop-dependencies.ts` installs the X server, window manager, dock, compositor, clipboard bridge, terminal, fonts and Chrome libraries through image-root `/usr/bin/apt-get` with `--no-upgrade` and `--no-remove`. The installer uses only system command paths, bypassing Kata persistent-apt wrappers so binaries, X assets and shared libraries live in the same filesystem. Packages must be installed again when a Kata save or container recreation discards that root; Chrome and its profile remain in persistent storage. Google Chrome is an exact-version, SHA-256-verified download for Linux x64 or ARM64, extracted under the assistant's internal external-dependency directory. Extraction does not run Chrome package scripts, register its repository or change the system's default browser. Successful setup is recorded only after Chrome runs; readiness also checks the desktop binaries and X fonts so a recreated container offers setup again. Installation errors remain retryable. The base image carries no desktop-only packages or baked browser; existing browser-tool and PDF installations retain their own Playwright behavior.
 
 `DesktopSessionManager` owns `Xtigervnc` on display `:99` with VNC on `localhost:5999`, `openbox`, `xcompmgr`, `tint2`, `tigervncconfig` and Google Chrome. Chrome starts directly, without Playwright or automation switches, using the existing `data/desktop-profile` directory. The dock configuration and launcher paths remain under `data/desktop-panel`. Children receive only an allowlisted environment. One viewer holds the slot at a time; the tree lingers five minutes after disconnect. Required child failures and browser crash loops tear down the tree, while cosmetic dock/compositor failures only log. Shutdown uses SIGTERM followed by SIGKILL after a two-second grace, and a subsequent start waits for teardown. The `assistant-desktop` flag and `IS_CONTAINERIZED` gate both setup and streaming. The companion platform PR's memory and shared-memory provisioning remains independent of package installation.
